third_party/blink/web_tests/tables/mozilla_expected_failures/bugs/bug3517-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: dump is not defined
